Move interests section into components folder
